create a context menu similar to the native triggered when right-clicking an anchor tag.
The problem with the native context menu is that it automatically hides the chrome extension popup after the user right clicks and selects open in tab. That didn't seem like the right user behavior to me especially that the newly opened tab is not focused/active. It also allows for marking as read only when the post is opened (not when copying the link only and clicking some other context menu action).